# src/lib/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CHALLECARA API Documentation",
    "version": "1.0.0",
    "description": "API documentation for CHALLECARA - A social profile management platform",
    "contact": {
      "name": "CHALLECARA Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://challecara.com",
      "description": "Production server"
    }
  ],
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique user identifier"
          },
          "accountId": {
            "type": "string",
            "minLength": 3,
            "maxLength": 20,
            "description": "User account ID"
          },
          "nickname": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "description": "User nickname"
          },
          "bio": {
            "type": "string",
            "description": "User biography"
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri",
            "description": "Avatar image URL"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Account creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": ["id", "accountId", "nickname", "bio", "avatarUrl", "createdAt", "updatedAt"]
      },
      "UserCreateInput": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "minLength": 3,
            "maxLength": 20,
            "description": "User account ID"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "format": "password",
            "description": "User password (minimum 8 characters)"
          },
          "nickname": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "description": "User nickname"
          },
          "bio": {
            "type": "string",
            "description": "User biography"
          },
          "avatarUrl": {
            "type": "string",
            "format": "uri",
            "description": "Avatar image URL"
          }
        },
        "required": ["accountId", "password", "nickname", "bio", "avatarUrl"]
      },
      "SocialLink": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique social link identifier"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "Associated user ID"
          },
          "provider": {
            "type": "string",
            "enum": ["twitter", "instagram", "facebook", "tiktok"],
            "description": "Social media provider"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Social media profile URL"
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the link is active"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          }
        },
        "required": ["id", "userId", "provider", "url", "isActive", "createdAt"]
      },
      "SocialLinkCreateInput": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "Associated user ID"
          },
          "provider": {
            "type": "string",
            "enum": ["twitter", "instagram", "facebook", "tiktok"],
            "description": "Social media provider"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Social media profile URL"
          }
        },
        "required": ["userId", "provider", "url"]
      },
      "BlogPost": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique blog post identifier"
          },
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "Author user ID"
          },
          "title": {
            "type": "string",
            "description": "Blog post title"
          },
          "content": {
            "type": "string",
            "description": "Blog post content"
          },
          "isPublished": {
            "type": "boolean",
            "description": "Whether the post is published"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": ["id", "userId", "title", "content", "isPublished", "createdAt", "updatedAt"]
      },
      "BlogPostCreateInput": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "format": "uuid",
            "description": "Author user ID"
          },
          "title": {
            "type": "string",
            "description": "Blog post title"
          },
          "content": {
            "type": "string",
            "description": "Blog post content"
          },
          "isPublished": {
            "type": "boolean",
            "default": false,
            "description": "Whether the post is published"
          }
        },
        "required": ["userId", "title", "content", "isPublished"]
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          },
          "code": {
            "type": "string",
            "description": "Error code"
          }
        },
        "required": ["error"]
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "tags": [
    {
      "name": "Users",
      "description": "User management endpoints"
    },
    {
      "name": "Social Links",
      "description": "Social media link management"
    },
    {
      "name": "Blog Posts",
      "description": "Blog post management"
    },
    {
      "name": "Authentication",
      "description": "Authentication endpoints"
    }
  ],
  "paths": {
    "/api/blog-posts": {
      "get": {
        "summary": "Get blog posts",
        "description": "Retrieve blog posts, optionally filtered by user ID",
        "tags": ["Blog Posts"],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by user ID"
          },
          {
            "in": "query",
            "name": "isPublished",
            "schema": {
              "type": "boolean"
            },
            "description": "Filter by published status"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of blog posts",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BlogPost"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a blog post",
        "description": "Create a new blog post",
        "tags": ["Blog Posts"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BlogPostCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Blog post created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BlogPost"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/openapi": {
      "get": {
        "summary": "Get OpenAPI specification",
        "description": "Returns the OpenAPI 3.0 specification for the CHALLECARA API",
        "tags": ["Documentation"],
        "responses": {
          "200": {
            "description": "OpenAPI specification",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/api/social-links": {
      "get": {
        "summary": "Get social links",
        "description": "Retrieve social links, optionally filtered by user ID",
        "tags": ["Social Links"],
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "schema": {
              "type": "string"
            },
            "description": "Filter by user ID"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of social links",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SocialLink"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a social link",
        "description": "Add a new social media link for a user",
        "tags": ["Social Links"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialLinkCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Social link created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SocialLink"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "summary": "Get user by ID",
        "description": "Retrieve a specific user by their ID",
        "tags": ["Users"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update user",
        "description": "Update an existing user's information",
        "tags": ["Users"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nickname": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 50
                  },
                  "bio": {
                    "type": "string"
                  },
                  "avatarUrl": {
                    "type": "string",
                    "format": "uri"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete user",
        "description": "Delete a user account",
        "tags": ["Users"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully"
          },
          "404": {
            "description": "User not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/users": {
      "get": {
        "summary": "Get all users",
        "description": "Retrieve a list of all users",
        "tags": ["Users"],
        "responses": {
          "200": {
            "description": "A list of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new user",
        "description": "Create a new user account",
        "tags": ["Users"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreateInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  }
}